--- internal/timebot/getCurrentUtcTime.go
@@ -1,7 +1,0 @@
-package timebot
-
-import "time"
-
-func GetCurrentUtcTime() string {
-	return time.Now().UTC().Format(time.RFC3339)
-}
