+++ internal/views/cards/timeUpdate.templ
@@ -1,7 +1,7 @@
 package cards
 
 import (
-	"github.com/jgndev/jgn.dev/internal/local"
+	"github.com/jgndev/jgn.dev/internal/timebot"
 )
 
 templ TimeUpdate() {
@@ -12,26 +12,26 @@ templ TimeUpdate() {
 			<div class="flex flex-col">
 				<!-- US Time Zones -->
 				<span class="text-4xl uppercase font-bold text-center">USA</span>
-				@TimeCard("Eastern (ET)", local.GetCurrentTime("America/New_York"), "et-time")
-				@TimeCard("Central (CT)", local.GetCurrentTime("America/Chicago"), "ct-time")
-				@TimeCard("Mountain (MT)", local.GetCurrentTime("America/Denver"), "mt-time")
-				@TimeCard("Pacific (PT)", local.GetCurrentTime("America/Los_Angeles"), "pt-time")
+				@TimeCard("Eastern (ET)", timebot.GetCurrentTime("America/New_York"), "et-time")
+				@TimeCard("Central (CT)", timebot.GetCurrentTime("America/Chicago"), "ct-time")
+				@TimeCard("Mountain (MT)", timebot.GetCurrentTime("America/Denver"), "mt-time")
+				@TimeCard("Pacific (PT)", timebot.GetCurrentTime("America/Los_Angeles"), "pt-time")
 				<!-- EU Time Zones -->
 			</div>
 			<div class="flex flex-col">
 				<span class="text-4xl uppercase font-bold text-center">EU</span>
-				@TimeCard("Western (WET)", local.GetCurrentTime("Europe/London"), "wet-time")
-				@TimeCard("Central (CET)", local.GetCurrentTime("Europe/Paris"), "cet-time")
-				@TimeCard("Eastern (EET)", local.GetCurrentTime("Europe/Bucharest"), "eet-time")
-				@TimeCard("Moscow", local.GetCurrentTime("Europe/Moscow"), "msk-time")
+				@TimeCard("Western (WET)", timebot.GetCurrentTime("Europe/London"), "wet-time")
+				@TimeCard("Central (CET)", timebot.GetCurrentTime("Europe/Paris"), "cet-time")
+				@TimeCard("Eastern (EET)", timebot.GetCurrentTime("Europe/Bucharest"), "eet-time")
+				@TimeCard("Moscow", timebot.GetCurrentTime("Europe/Moscow"), "msk-time")
 			</div>
 			<!-- APAC -->
 			<div class="flex flex-col">
 				<span class="text-4xl uppercase font-bold text-center">APAC</span>
-				@TimeCard("India", local.GetCurrentTime("Asia/Kolkata"), "ist-time")
-				@TimeCard("Hong Kong", local.GetCurrentTime("Asia/Shanghai"), "cst-time")
-				@TimeCard("Tokyo", local.GetCurrentTime("Asia/Tokyo"), "jst-time")
-				@TimeCard("Sydney", local.GetCurrentTime("Australia/Sydney"), "aedt-time")
+				@TimeCard("India", timebot.GetCurrentTime("Asia/Kolkata"), "ist-time")
+				@TimeCard("Hong Kong", timebot.GetCurrentTime("Asia/Shanghai"), "cst-time")
+				@TimeCard("Tokyo", timebot.GetCurrentTime("Asia/Tokyo"), "jst-time")
+				@TimeCard("Sydney", timebot.GetCurrentTime("Australia/Sydney"), "aedt-time")
 			</div>
 		</div>
 	</div>
