--- internal/views/lockups/timeBot.templ
@@ -1,24 +1,0 @@
-package lockups
-
-import (
-	"github.com/jgndev/jgn.dev/internal/views/icons"
-	"github.com/jgndev/jgn.dev/internal/views/cards"
-)
-
-templ TimeBot() {
-	<div class="max-w-4xl mx-auto">
-		<div class="my-4 flex items-center justify-center">
-			<div class="flex flex-col items-center">
-				@icons.TimeBotIcon()
-				<span id="copy-tooltip" class="pt-2" style="visibility:hidden;">Copied!</span>
-			</div>
-		</div>
-		<div
- 			hx-get="/get-time"
- 			hx-trigger="every 1s"
- 			hx-swap="innerHTML"
-		>
-			@cards.TimeUpdate()
-		</div>
-	</div>
-}
