+++ internal/views/pages/timeBot.templ
@@ -2,11 +2,26 @@ package pages
 
 import (
 	"github.com/jgndev/jgn.dev/internal/views/shared"
-	"github.com/jgndev/jgn.dev/internal/views/lockups"
+	"github.com/jgndev/jgn.dev/internal/views/cards"
+	"github.com/jgndev/jgn.dev/internal/views/icons"
 )
 
 templ TimeBot() {
 	@shared.Base("Time Bot") {
-		@lockups.TimeBot()
+		<div class="max-w-4xl mx-auto">
+    		<div class="my-4 flex items-center justify-center">
+    			<div class="flex flex-col items-center">
+    				@icons.TimeBotIcon()
+    				<span id="copy-tooltip" class="pt-2" style="visibility:hidden;">Copied!</span>
+    			</div>
+    		</div>
+    		<div
+     			hx-get="/times"
+     			hx-trigger="every 1s"
+     			hx-swap="innerHTML"
+    		>
+    			@cards.TimeUpdate()
+    		</div>
+    	</div>
 	}
 }
